Update 08-03-18
These are the updates I have so far. I've added the states outline to layer control and have the colors defined by whether they have removed monuments (yellow) or not. All stylistic changes can of course be modified. I did not add the monuments to the layer control because we may want them to stay on there the whole time and the way I added the states does not use the ajax call, so I didn't want to mess with that piec e knowing edits were being made to the style of the points. Also, the story map includes points for each area it discusses, and those can be removed or the style changed.